requirements/local.txt
-r ./base.txt
Werkzeug==1.0.1 
ipdb==0.13.2  
Sphinx==3.0.1  
psycopg2==2.8.5 --no-binary psycopg2  

mypy==0.770  
django-stubs==1.5.0  
pytest==5.3.5  
pytest-sugar==0.9.2  

flake8==3.7.9  
flake8-isort==3.0.0  
coverage==5.1  
black==19.10b0  
pylint-django==2.0.15  
pre-commit==2.2.0  

factory-boy==2.12.0  
django-debug-toolbar==2.2  
django-extensions==2.2.9  
django-coverage-plugin==1.8.0  
pytest-django==3.9.0
drf-yasg==1.17.1
